Fix hanging test.
The problem was encode; some kind of deadlock interacting with the capnp
tool. The way we were doing this was dumb anyway, and mostly a product
with me not being terribly familiar with the System.Process API. The
major change is we're no longer plumbing the data from the .capnp
constant file into capnp compile ourselves; just hand it the file
descriptor.