Untyped: Remove indrection in `index`.
Rather that returning a pointer to the value, it now just returns the
value. This was a remnant from when I was thinking I'd use this module
for writing as well as reading.

I've been struggling to figure out how to get the schema stuff working,
and the indirection was making things harder, since it introduces an
impedance mismatch between how I want the typed api to look and how the
untyped api looks.

In addition, there were a few conceptual problems with it, It wouldn't
necessarily have been valid to write a `Struct b` to a `PtrTo b (Struct
b)`, since struct sizes vary.