More thinking out loud re: typed readers
At this point I'm thinking of just having readers & builders be entirely
separate, and rely on namespacing to keep them separate. This would also
mean we could get rid of the (%~) and (<~) operators, and just make e.g.
the builder version of displayName be of type Text -> Builder Node ...
instead of Field Node Text.