Work towards getting new test to pass
The pattern match failure was because I'd forgotten to add a case for
ListOfMapped when I added that variant. Once this was solved it revealed
a bug -- my TH generator for list readers always used pointer 0; a
mistake when generalizing it from the hand-written nodes reader.

This is still failing; it looks like capnp compile generated a far
pointer, which we don't handle right yet. I guess we need to do that.