Index blobs by byte, not word.
The main motivation for this is so we can more conveniently work at the
byte level where it makes sense. In particular, it would be nice to just
expose capnproto's Text and Data as BlobSlices into the underlying
storage, but this is awkward if we're doing word-indexing.

In practice all the implementations were byte-oriented under the hood
anyway.