Add rawBytes function.
Not exported yet; putting this down because I'm not really in the right
mental shape to be hacking right now.

This is intended to be used for extracting Data and Text types. It
arguably makes more sense to define it for all list types, to avoid the
weird exception behavior, but right now I can't be bothered; see above.