Starting making use of stylish-haskell.
Adds a `make format` target to apply to every .hs file in the repo, and
a config file that's not a no-op. The rest of the diff is the result of
running `make format`.

Right now, Language.CapNProto.TH fails to parse; stylish-haskell doesn't
seem to like the one of the pattern splices. I'll need to figure out
what to do with that.