More import cleanup.
As part of this, I noticed somewhere where I had:

    import Data.CapNProto.Blob as B

Which was meant to be qualified but wasn't. This corrected now, and the
few places that were using un-qualified names have been fixed to match.

There's a method to the way I've re-organized, though the goal is mostly
to have some kind of convention; will document soon.