Don't use quota in Untyped.
This is part of an experiment to make the API more reasonable; without
the quotas we can use ViewPatterns & a concrete instance of MonadThrow
(e.g. Maybe) to get our pattern matching back.

We'll need to do something else about the DoS attacks this opens up;
still thinking about that. Maybe something with a timeout.