Don't wrap list pointers in Maybe
...just return a empty list.

We'll do this for other pointer types soon too. Note that this is
currently breaking the testsuite's build. The test suite was relying
on the Maybe to determine if there was an annotation list or not, but
now we have no way of doing that. The plan is to add hasFoo for pointer
fields, much like other libraries do.