Simplify/clean up untyped tests.
It seems unlikely that I'll ever add a second item to either of these
lists, so the `tst` helper function is superfluous. We could get rid of
the list itself as well, but we'd have to change assertionsToTest, which
might be desirable, but I'm not doing it just now.

I've also factored out a bit of the common encoding logic.

We could factor out the value that we're using as well, but I think the
tests are easier to read with the value in-line.