hlint: don't check top-level directory.
I've been doing some ad-hoc testing running the compiler plugin and
writing the output to a file in $PWD (the root of the repo). But hlint
complains about the generated code, and I run it in a pre-commit hook,
so I have to remember to delete the file before committing. This patch
solves this by restricting the checks to directories where code is
actually supposed to live.

I also delete the reference to the now-absent directory for the old code
generator's output.