Replace the hand-written data definitions for schema.
...with the generated ones. The Decerialize instances are still hand
written, plus a couple other things.

The code generator is using this, and outputs the same code still. Tests
have been adjusted and still pass.

A few tweaks needed to be made to the instances, mostly due to me not
having written the fields in quite the same order as the code generator.