Split HsAst up.
...between the data types and the formatting.

Per the comment at the top of HsSchema, this is going to be a slightly
more intermediate representation; I'd originally conceived of this as a
simplified Haskell AST with just the features we actually need (code
that uses e.g. Template Haskell can be very verbose due to the
generality of the data types).