Move pointer extraction logic to a type class.
This will make it easier to generate the corresponding logic for
non-primitive types, as we can generate the code at the type definition
site, where we have the information.

Note that since this is no longer conditionally spitting out
`undefined`, the generated schema do not currently build; as such, the
modules have been commented out in capnp.cabal.