*Actually* re-enable old tests and get them passing.
I'd uncommented the modules in the cabal file, but not actually added
the tests to the test suite's main.

Doing so revealed that the has_* helpers did not have quite the
semantics I intended; for pointer fields, an null pointer should count
as False, not just if the pointer section is too small.