Rework regen.sh
The biggest difference is that we no longer copy the schema from the
system, but generate codec directly from the copies we already have
locally.

To facilitate this, the schema need to be inside a directory called
`capnp`, because they import one another with paths starting with
"/capnp". Accordingly, we've moved the *.capnp files to
`lib/schema/capnp`.

Finally, we run the code generator from within lib/, so we don't have
to then move the generated output.