Add copyright line to cabal file
...and a CONTRIBUTORS file.

Note that while contributors do not necessarily own the copyright to
their work (employers tend to if the contributor works professionally as
a developer), I think at present everyone on that list did own the
copyrights when they submitted -- besides myself (working as a
freelancer), Andrew and Remy were students at the time. I should verify
this before I explicitly document it though.