Remove the MMessage type class.
We only have one instance of this, and the fact that the instance is
called "MutMessage" is a red flag that this is probably unnecessary
abstraction. I can think of some use cases for other instances (e.g.
stuff that can't be memory mapped), but I'm going to lean in favor of
simplifying the API.