Cleanup in the representation of messages.
They're now actually vectors of *segments*, rather than vectors of
vectors. The runtime representation is actually the same, but this
is conceptually cleaner.

We couldn't do this before when Segment had the 'm' type parameter,
so the old representation was a bit of a hack. The 'm' parameter was
only needed for the MMessage typeclass, which has been removed -- so
we can simplify further now.