Fix bug allocCompositeList.
This was the cause of the test failures. I'd been working from memory
when writing this code, and mis-remembered the details of the composite
list pointer encoding. The offset in the tag word is the length in
elements (while the length field in the pointer is the length in words).
I'd for some reason been thinking the offset was just unused :/.