Make modDecls a Map.
I'm going to want to look up definitions of other type from
fmtFieldAccessor when generating setters for unions, so this is a step
towards that. Per the comment, we should do a bit of refactoring; the
ModuleRef doesn't need to be part of the map key.

Note that this causes a lot of semantically meaningless churn in the
generated output, because the order of the declarations is changed.