Re-introduce struct wrapper around unions in raw backend.
For the pure backend, we translate:

    struct Foo {
        union {
            bar @0 :Bool;
            baz @1 :Int32;
        }
    }

to:

    data Foo = Bar Bool | Baz Int32

i.e. we remove the outermost struct wrapper, as it's mostly noise.

However, for the raw backend, this prevents us from modifying a Foo
in-place, since the Bar/Baz tag don't provide access to the underlying
struct.

The pure backend generates the same data structures as before, and is
only changed to handle updates to the raw backend.

The raw backend now generates a struct wrapper even in the case of a
struct that is one big anonymous union.