Don't define enums twice.
The definitions in Raw vs Pure modules is exactly the same -- so we just
re-export the Raw ones in the Pure modules.

This also adds deriving(Show, Read, Eq) to the enums in the raw modules.
By virtue of being the same type, the Pure enums now implement Enum,
IsWord, and other type classes.