Generate marshaling code for Data fields.
The only place in the standard schema where this actually shows up is in
the data variant of Value in schema.capnp -- which has a variantParams
of Unnamed, so we're not currently generating code for that anyway :/.
So this is untested. We'll see how it goes once we implement that.