Start playing with refactoring the IR.
There are a whole bunch of illegal structures that are nonetheless
representable in the IR, and we have a whole bunch of partial functions
in the compiler plugin because of it. This patch marks the beginning of
an effort to fix that.

This doesn't build, and right now my priority is getting write support
to MVP, so I'm pushing this to a branch and may revisit later.