Don't generate a call to decerialize for types that don't need it.
For fields that go in the data section (and void fields), this is always
a no-op, and doesn't even change the type, so we can just omit it.

Leaving it in simplifies the code generator, but my plan is to collapse
the Cerialize and Decerialize classes, and these basic types can't
correctly implement marshalInto from Cerialize, since they don't hold
any reference to the message. This patch makes it so we don't need them
to implement Decerialize either, so that's fine.