Fix test failure.
Define equality for (pure) structs more loosely, so that the sections
don't need to be the same size, so long as the bigger one is padded
with default values.

This was the source of the test failure; cerializing and decerializing
causes a list of structs to all take the maximum section sizes.