Add a big pile of comments.
Almost every exported name is documented now (not counting the generated
code).

Additionally, this commit adds a module Data.Capnp.Tutorial. Right now
this is just the text of the readme; I plan for the two to diverge, with
this module taking over instructional material, and of course changing
the formatting from markdown to haddock markup.