Refactor the code generator plugin.
The idea here is to draw a stronger distinction between
unions/groups/structs/other data type definitions. Not
having this is making codegen for interfaces weird.

Unfortunately the fact that we're flattening
struct { union { ... } } in the *frontend* means we lose
access to the outer struct's metadata, so that still needs
to be re-jiggered.