Start playing with an API for accessing capabilities
The thing I'm trying to address here: when traversing a message, when we
hit a capability pointer, it's just a Word32 -- how do we correlate that
to the actual capability? This information has to be provided out of
band. In the rpc protocol, payloads include a capTable field that can be
used to derive this information, but fundamentally the info isn't *in*
the message.

The Go implementation (the other implementation that I am most familiar
with) just has a client table as part of the message object, but this is
unattractive for a few reasons:

* It makes module dependencies a mess; we'd need to define clients in
  Data.Capnp.Message to avoid a cyclic dependency, and then if we want
  to keep it abstract we basically need to fold the rest of the RPC
  system into the same module.
* It precludes alternate implementations of clients, unless we're going
  to add another type parameter to our message types/type classes, and
  I'd like to avoid that.
* It's just conceptually weird; the cap table really *isn't* part of
  the message.

My notion is that we'll have the ability to lookup caps be part of an
mtl-style type class. To extract a cap from a message, you need to be
in such an instance.

NoCapT can be used when reading messages when no capability table is
available. There, you just get the raw Word32s, which you can't really
do anything with.

We'll also add implementations of MonadGetCap that are essentially
ReaderT around the cap table, and can be used when the RPC system is
actually available.

This is still just an exploration; the design might change completely
by the time something is working.