Change toPtr to be monadic.
We need this because instances for interfaces will need to modify the
capability table.

Keeping this building meant having to (1) run the code generator, whose
output had been out of sync due to build failures from
partially-implemented interface support, and (2) avoid building
Persistent, which was the only schema actually affected.