Start working on issue #34.
We have a failing test case (though we should expand it at some point to
be more thorough), and a solution that doesn't build.

The build errors are because I tried to make it possible to copy data
from different message *types*, but the equality test makes this not
feasible.