Fix performance problem.
We were allocating stupid amounts of memory given the old implementation
of socketTransport, because:

1. We allocate a buffer with room to spare.
2. We read in the whole message at once, which turns out to be **much**
   smaller than the buffer.
3. We have lots of unused space in the message we're passing around.
4. We have lots of garbage to clean up later.

I've just deleted the old implementation in favor of doing the same
thing that handleTransport does, which seems to perform reasonably.
We have to check for short reads ourselves, but that's the only thing.
Some related functionality has been changed/removed accordingly.

Ironically, it was written this way to make smarter use of syscalls. Yay
premature optimization?