Add a private 'aircraftlib' library.
I plan to use this in the test suite. Part of why the output isn't just
in the test suite proper is that tests/tests/Capnp/Gen is a symlink to
examples/Capnp/Gen, for the echo example. This is a bit awkward, but I
really want to keep the examples directory relatively self-contained.

Due to #28, I've had to comment out a few things in the schema to
satisfy code generator; see the comment at the top of the file.

Finally, the aircraft modules do not actually build currently (and are
commented out to prevent them from breaking the build as a whole),
because there is a bug (which this has unearthed) where a struct with no
fields does not result in a type definition. Need to fix.