Remove CommitThrow.
This was why we weren't sending the abort; I'd misunderstood how
throwSTM and catchSTM interact, and so it was still rolling back the
transaction. See also:

    https://github.com/ghc/ghc/pull/210

We're still getting a test failure -- a TraversalLimitError when
decoding what should be the abort message. This is weird; we're setting
the limit to maxBound, which should be (2^63-1), so even if it's cyclic
data it shouldn't be hitting that in a reasonable amount of time. Needs
further debugging.