Fix #56
In most places, we are calling invoice with word count, but in rawBytes
we were calling it with a byte count. Combined with the fact that
bsToValue sets a limit based on the message size (and therefore doesn't
need to be run inside of LimitT, and ignores the limit), this meant that
we were hitting the limit when decoding messages that were mostly one
long string.