Add a size hint parameter to newMessage.
The primary reason for this is that we'll need this to implement
canonicalization (#36), in light of #57, since the default will no
longer always allocate inside the first segment.

This also reverts the change to the test suite made in #57; we do
want to verify that the new segment has the index we think it does,
and adding a larger default segment size has the effect of no-longer
breaking that test.