Don't allocate question/export ids in their own transactions.
This was initially done as part of a misguided premature optimization:
I'd been thinking the allocation pools might be a point of contention,
whereas we can use things from stm-containers so that most calls and
such don't touch the same tvars and thus don't block each other. But:

1. None of this has been profiled
2. We have other things (e.g.  the rate limiting semaphores) that have
   the same issues anyway.
3. Most importantly, I worry about what happens if a thread gets
   cancelled between allocating an id and using it -- we could
   permanently reduce the effective size of the id pool.

   This could even become remotely-triggerable if e.g. we decide to allow
   cancelling an outstanding call by sending a finish message.

So for now we do the simple, and more clearly correct thing -- allocate
ids in the same transaction. We can revisit optimizations after actually
profiling.