Start experimenting with re-organizing the rpc implementation.
It's gotten too unwieldy for me to work with; I can't keep all of the
interactions in my head at once, so I plan on reorganizing it. I'm not
throwing out the old implementation, but will be copying things over
piecewise. The Note [Organization] in the new source points to the main
thing I want to change; the old implementation has a bunch of logic in
the coordinator, and so related logic ends up strewn about the module.