Remove incRef/decRef, for now.
The garbage collector will already do the right thing here, and the
initial create of a LocalClient is going to keep an extra reference
floating around anyway, so we already won't get prompt release.

These never really made a ton of sense for the other client types
anyway. Now that all of those TODOs are gone, the tests are passing
again.