Refactoring re: nullClient.
We had an error "TODO" that was only there because the function
conceptually shouldn't ever be handed a null client. In fact I believe
this could happen if a vat was given a null bootstrap interface (as
opposed to *no* bootstrap interface).

This patch refactors the internals, so that it is possible to talk of a
non-nil client as a type. It also fixes the above issue re: bootstrap
interfaces.