waitForServer: be less generous.
This was long enough that with a handful of tests using it (sometimes
twice), It was a bit annoying to wait for. The initial value was chosen
when we were invoking cabal-new run from within the tests, which has a
long startup time, so we can get away with something shorter now.