Rework the FinalizerKey abstraction.
It's now a cell which contains a value, so we can more closely associate
it with the thing that it logically is attached to.

Note that the code using finalizer keys has not been adjusted to really
make use of this property, yet.

Also, it is possible to add more finalizers without bumping the old
ones. I don't plan on making use of this, but it's a safety thing.