Start doing a bit of readability cleanup in the code generator.
We've got a lot of places in the code generator where there are these
huge functions. From an architectural standpoint they're fine, and
splitting them up isn't a big deal in terms of separation of concerns,
but just having them in one block like that makes them harder to read.
Lesson learned. I plan on doing this to much of the code generator, bit
by bit. Trans.PureToHaskell is first since that's what I'm otherwise
working on right now.