Process whole CGRs as a unit.
This makes things a little more robust, but we have a fundamental
problem: if a node is defined in a file not in requestedFiles, it may
be referenced but not actually in the cgr in the first place. This means
needing to rethink how we generate instances of interfaces classes.