Be smarter about what imports we generate.
Rather than trying to figure it out from the input, we just generate the
haskell module, and then scan it to figure out what imports we need.

This was necessary because of the following scenario:

* a.capnp defines an interface IA
* b.capnp defines an interface IB, which inherits from IA. b.capnp imports a.capnp
* c.capnp defines an interface IC, which inherits from IC. c.capnp imports b.capnp

Previously in the module output for c.capnp, we'd generate a newtype
wrapper around a client for `IC`, and generate *'server_ instances for
`IA`, `IB`, and `IC` -- but it won't actually import the class for `IA`,
so the result won't build.

Previously it was ok to just look at the imports in the
requestedFile.imports field, but we may actually have to directly import
transitive dependencies now.

Currently the fix is only used in the high-level modules, since that's
where it's needed, but soon it will be used everywhere for consistency.

In the core schema. This actually has the effect of removing some
imports that were unused (c++.capnp).