Use string literals for bytestrings constants
..rather than BS.pack [0,1,2,...]

This makes a *dramatic* difference on both the (haskell source) size and
(ghc) build time of the modules generated from the sandstorm schema.

One particularly pathological example: sandstorm's mime.capnp is
dominated by one giant table in the form of a constant declaration,
with both keys and values being Text. The source file is about 56KiB.
The size of the output, with the old version of the code generator, was
almost 4MiB! The new output size is 196KiB, which is much more
reasonable. Build time for that file is now quite fast, whereas before
it was one of the slowest ones.

Part of the problem is the way we were pretty-printing things, which
could have been solved in other ways, but this will also improve the
size of constants that are mostly text, even without that. I also
suspect this will improve build times more than just fixing the
pretty-printing would.