Fix #71
In the case that I hit the missing node was 'SourceInfo', which we don't
actually use. In general, nodes from other schema can get pulled in as
dependencies without transitively pulling in all of their NestedNodes,
so we need to handle that scenario gracefully. In this case we can just
omit the missing nodes from the map, as they aren't used.