Update core-schema to 0.8
This required some mechanical code changes in the RPC layer now that
CapDescriptor is no longer of the form `struct { union { ... } }`,
and thus the code generator no longer omits the wrapper around the
union. We don't care about the new field currently, so it's just a
matter of wrapping and unwrapping.