Add an extra MutVar aroudn the AppendVec in MutSegment.
This solves the last of the test failures; optimizing by hanging on to a
reference to the segment actually breaks the old implementation, since
the allocated length is not updated in-place. This was manifesting as a
spurrious bounds check failure.

Hopefully the overhead of an extra indirection reading a MutVar will
still be faster than reading the segment out of the segments array
(including the associated bounds check).

If not, there is room for further optimization here; we no longer
actually grow individual segments, so we could drop some of the code
that manages that to make things faster, and just have a direct
reference to the vector. But let's measure first.