More optimization.
Mostly optimizing Message.write.

This started off as an attempt to remove some of the cruft in the
allocation routines, get rid of the now-superfluous AppendVec for
segments, and remove the indirection of the MutVar for access to the
segment data (since we never actually resize the segments anymore).

In the course of doing this I discovered we actually *do* still
sometimes resize segments, specifically in the case where a double-far
pointer could be used. This was probably bad for performance, though
I'm not sure how often it came up in practice.

This patch gives us another ~20% speedup for ocap-merkledag.