Avoid extra marshal of call parameters
This gives us about a 2x speedup on ocap-md.

We're still doing some unnecessary *copies* in various places, which is
probably the next thing to work on re: rpc perf. But this also probably
requires tweaking some APIs; the copies are there because our hand is
forced due to the need to thaw in some places.