Add some extra bounds checking when reading a pointer.
This isn't currently strictly necessary for security, since accesses to
the underlying message is also bounds checked, but it will become so
soon, due to an optimization I want to make.

Specifically, Untyped.setIndex has proven to be a bottleneck when
copying & canonicalizing data in ocap-merkledag, where much of the data
is just flat lists of bytes. So, I want to make it so that copying a
list of bytes just calls libc's memcpy(), which should be very well
tuned and faster than calling index/setIndex. But since memcpy doesn't
do bounds checking, we need to make sure that we do the checks ahead of
time.

Before actually doing this optimization, we need to add tests for
messages that actually have out of bounds pointers.