Add an INLINE pragma for replaceBits.
While profiling ocap-merkeldag, I noticed that some amount of allocation
was attributable to replaceBits -- meaning the type class dictionaries
were not being inlined. This is a low level function doing bitwhacking
and that's going to have a noticeable impact on performance; let's give
the compiler a hint.

If I were to do it again I'd probably making the bitwhacking more
straightforward.