Struct Allocate instances, tests.
This adds a new class 'TypedStruct' to be implemented by each generated
struct; it provides the sizes of the struct's sections, which is used
by the 'Allocate' instances.

The tests currently don't build due to an ambiguous type; it seems
initVariant doesn't give the type checker enough of a hint. Need to
think about how to improve this.