Fix type inference problems.
Rework the relation ship between IsLabel and HasField (and HasVariant),
such that using overloaded labels for fields/variants is likely to work
better. The test case typechecks without annotations now (except the one
for newRoot).

The test itself fails; it looks like the value that is coming out is the
default value. Need to investigate.