Fix a codegen bug re: union tags.
tagOffset originates as disriminantOffset in the schema, which is in
units of 16-bits, but dataOff is supposed to be in units of bits. So we
need to multiply.

This whole thing is a bit messy; I had to actually track tagOffset
through the program to remind myself where it came from. Should document
the IR, and maybe define a 'Mutlple' type for fields like this.

The tests are passing again.