Let's not leave too much uncommitted.
This started as a patch to port the code generator to use the repr
APIs, largely as a way to test them. Doing this found a couple issues,
Some of which I fixed in stand-alone commits before this one.

The current issue is that there was no implementation of Parse.encode
for Lists. Adding this has turned into a pie of fiddly type class
meta-programming. This is not a great stopping point, but I'm very wary
of keeping a bunch of state uncommitted, so here we are.