Fix a double-finish bug.
We were sending extra finish messages for PromiseClients that point at
an a PromisedAnswer created via a Call message.

For PromiseClients created in call messages, we don't need to release
them specifically, since the finish message is sent by cbCallReturn.
All that remains is to send finish messages for bootstrap messages,
which we add to this patch as well.