Get rid of `cpu` package; use stuff from base instead.
This improves the benchmark perf by ~5%, probably due to avoiding the
FFI. Previously, fromLE64 was showing up in the profiling output as a
source of a couple percent of allocations, which was a red flag -- it's
a no-op on my cpu, so it *really* shouldn't be allocating.

Removing a dependency is also just always nice.