Fix a DoS vulnerability I introduced since the last release.
This never made it into a release, but I realized it was here and it's
been on my TODO list. Release is immanent, so it's time.

This was introduced when I moved the invoicing to happen on pointer
dereference, rather than on word read. But we still need to make sure we
don't endlessly follow far pointers. Per the comment, we should just
not do this by naively recursing.

One of the tests needed updating due to the increased invoicing.