Port the RPC system to the new serialization API.
The only functional change is that now (Parsed Capnp.Gen.Capnp.Rpc.New.Exception)
is used in APIs instead of Capnp.Gen.Capnp.Rpc.Pure.Exception. But we no
longer have internal dependencies on the old API, so we can begin
slashing and burning (though we'll have to port the tests as we go).