Tweaks to Untyped/HasMessage/MessageDefault
...and much progress towards getting this to build -- errors are only in
the tests now.

We've flipped the arguments for Untyped, so that it matches Raw, and
made it return kind (Mutability -> *) (which is different in that it
allows us to define instances on @Untyped a@).

This has the annoying affect that we can't just do Untyped mut
('Data sz) = UntypedData sz, since the latter doesn't take a
mutability parameter. So there's an 'Unwrapped' type family that
gets used in some places.

I also tweaked HasMessage and MessageDefault to be defined on
`Mutability -> *`, which has been a mixed success -- it seems
to be really bad at type inference. Should probably define it on
the repr instead and take an Unwrapped (Untyped r mut).

The instance MaybeMutable (Raw a) is not yet defined, and it's
going to be slightly fiddly. I have a rough plan though.