Fix a race condition in callbacksLoop
Per the FIXME. Also, there was another of these on connection cleanup,
since we were just doing the flushCallbacks as a separate statement in
the do block.

Also also, this changes the logic from using sequence_ to chaining
things together with finally, which should ensure that later callbacks
will still run if an earlier one throws.

This should all be properly async exception safe now.