subscribeReturn: Don't defer Return callbacks to the cb loop.
Doing this resulted in a race condition:

Suppose alice lives in vat A, and bob and carol live in vat B

alice runs:

    carol = bob.getCarol()
    carol.foo()
    carol.bar()

With the old implementation it is possible for carol to receive bar() before foo();
here is the sceanrio:

- Bob sends the getCarol() message, and then immediately pipelines foo()
  onto its result, so the second call with have a messageTarget that is
  a PromisedAnswer.
- The server registers an onReturn callback for the getCarol() question,
  which will queue a call to foo().
- The call to getCarol() resolves. The onReturn call is put in the
  callbacks queue, and the return message is sent to vat A.
- vat B receives the call message for the call to bar(). Since the
  call to getCarol() had returned, this message has a messageTarget that
  is an importedCap, and points directly at the final destination. This
  call arrives before the callback loop has processed the onReturn
  callback, so it is delivered to carol before the call to foo().

The timing for this is obviously very sensistive, and unlikely to
trigger on a network with any meaningful latency. We were seeing it
in one of the tests that uses socketpair(), but only after removing some
of the queues that had previously been present.

As it turns out, nothing we actually put in subscribeReturn needs
deferral anyway, so let's just run it in the same transaction.